atomberg_sov/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
google-search-results==2.4.2
instagrapi==2.0.0
tweepy==4.14.0
textblob==0.17.1
nltk==3.8.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
python-dotenv==1.0.0
reportlab==4.0.7
pillow==10.1.0
